+++ src/index.ts
@@ -2,4 +2,4 @@ export { KnoraApiConnection } from "./knora-api-connection";
 export { KnoraApiConfig } from "./knora-api-config";
 
 // Classes
-export * from "./classes";
+export * from "./models";
